+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Set `updated_at` whenever it is saved
 - Updating a unique combatant updates their data throughout the application
 - Start/stop combat by clicking the initiative icon in the central table
+- Next/previous combatant by buttons or N/AltN
+- Combat round tracking
 
 ##[0.4.0]
 - Further implement Vuex, major refactor of use of variables and saving of state